pyproject.toml
[project]
name = "market-oppurtunity-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai==0.8.5",
    "langchain==1.0.2",
    "langchain-google-genai==3.0.0",
    "langgraph==1.0.1",
    "langsmith>=0.4.38",
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.22.1",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.15.0",
]
